Stub `iostream` (create `std::__msl_ios_base_{ninit,winit}`)
Several files include `iostream`, and so this gives us a nice
percentage boost. :^)

There is a problem, though: ninit and winit are constructed in
each file in the right order (ninit, then winit), BUT when they're put
in the .sbss section, they need to appear in the *opposite* order (so
winit, then ninit). `-inline deferred` may be causing this in some
way(?), but for now, this isn't something I know how to solve.

Once this is solved, we can mark some more files as Matching.