config: Put input code into a separate library
While working on CInput, I noticed that file had RTTI turned on, but
a bunch of other files in the engine don't. Also, it may not be using
`-inline deferred`, so there were probably different libraries with
different compiler flags.

This also gives me a reason to start organizing the engine folder. :^)