Initial import of THP from pikmin2 & prime
This is not completely a plug-and-play solution, some functions will
have to be matched by us (or we'll have to wait for someone else to
do it.)

This version of THP was built in late August 2002, except maybe not?
Five THP files use a different version of Metrowerks (assuming 1.3.2
for now) and different flags. THPDec, the file that contains the
version string, uses the older compiler & flags.

There are three games I know of that have THP decompiled:
  - smg1 (built in August 2007, five years difference)
  - pikmin2 (built in January 2004, much closer, but still quite a gap)
  - prime (same build date!! but... prime doesn't use as much of THP as
           Piglet does, only 2 files)

There's also a bss issue to deal with. May or not be related to #1.